Move database models to common
